third_party/WebKit/LayoutTests/http/tests/xmlhttprequest/origin-whitelisting-ip-addresses-with-subdomains-expected.txt

CONSOLE ERROR: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/get.txt. No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://localhost:8000' is therefore not allowed access.
CONSOLE ERROR: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/get.txt. No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://localhost:8000' is therefore not allowed access.
Specifying that an IP address should match subdomains doesn't make sense. This test verifies that it doesn't do anything.

Testing: http://127.0.0.1:8000/xmlhttprequest/resources/get.txt (sync)
FAIL: NetworkError: Failed to execute 'send' on 'XMLHttpRequest': Failed to load 'http://127.0.0.1:8000/xmlhttprequest/resources/get.txt'.
Testing: http://127.0.0.1:8000/xmlhttprequest/resources/get.txt (async)
FAIL: 0
